Move Lights model from Viewer to Renderer V2.
